split out to more separate classes
